pyproject.toml
[project]
name = "algovox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "edge-tts>=7.2.7",
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "kokoro-onnx>=0.5.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.41.0",
    "watchdog>=6.0.0",
    "webrtcvad-wheels>=2.0.14",
    "websockets>=16.0",
]
